chore : 도메인 별 패키지 분리, ctrl+alt+L, ctrl+alt+O
