+++ src/main/java/com/tourranger/airline/repository/AirlineRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.tourranger.airline.repository;
+
+public interface AirlineRepositoryCustom {
+}
