+++ src/main/java/com/tourranger/airline/repository/AirlineRepositoryImpl.java
@@ -1,0 +1,13 @@
+package com.tourranger.airline.repository;
+
+import org.springframework.stereotype.Repository;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+
+import lombok.RequiredArgsConstructor;
+
+@Repository
+@RequiredArgsConstructor
+public class AirlineRepositoryImpl implements AirlineRepositoryCustom {
+	private final JPAQueryFactory jpaQueryFactory;
+}
