--- src/main/java/com/tourranger/image/airline/dto/AirlineRequestDto.java
@@ -1,14 +1,0 @@
-package com.tourranger.image.airline.dto;
-
-import org.springframework.web.multipart.MultipartFile;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class AirlineRequestDto {
-	private String name;
-	private MultipartFile multipartFile;
-	private String url;
-}
