--- src/main/java/com/tourranger/image/airline/repository/AirlineRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.tourranger.image.airline.repository;
-
-public interface AirlineRepositoryCustom {
-}
