--- src/main/java/com/tourranger/image/airline/repository/AirlineRepositoryImpl.java
@@ -1,13 +1,0 @@
-package com.tourranger.image.airline.repository;
-
-import org.springframework.stereotype.Repository;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-
-import lombok.RequiredArgsConstructor;
-
-@Repository
-@RequiredArgsConstructor
-public class AirlineRepositoryImpl implements AirlineRepositoryCustom {
-	private final JPAQueryFactory jpaQueryFactory;
-}
