--- src/main/java/com/tourranger/image/thumbnailImage/dto/ThumbnailImageRequestDto.java
@@ -1,13 +1,0 @@
-package com.tourranger.image.thumbnailImage.dto;
-
-import org.springframework.web.multipart.MultipartFile;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ThumbnailImageRequestDto {
-	private MultipartFile multipartFile;
-	private String url;
-}
