--- src/main/java/com/tourranger/image/thumbnailImage/repository/ThumbnailImageRepositoryCustom.java
@@ -1,5 +1,0 @@
-package com.tourranger.image.thumbnailImage.repository;
-
-public interface ThumbnailImageRepositoryCustom {
-
-}
