--- src/main/java/com/tourranger/image/thumbnailImage/repository/ThumbnailImageRepositoryImpl.java
@@ -1,14 +1,0 @@
-package com.tourranger.image.thumbnailImage.repository;
-
-import org.springframework.stereotype.Repository;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-
-import lombok.RequiredArgsConstructor;
-
-@Repository
-@RequiredArgsConstructor
-public class ThumbnailImageRepositoryImpl implements ThumbnailImageRepositoryCustom {
-	private final JPAQueryFactory jpaQueryFactory;
-
-}
