--- src/main/java/com/tourranger/image/travelAgency/dto/TravelAgencyRequestDto.java
@@ -1,14 +1,0 @@
-package com.tourranger.image.travelAgency.dto;
-
-import org.springframework.web.multipart.MultipartFile;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class TravelAgencyRequestDto {
-	private String name;
-	private MultipartFile multipartFile;
-	private String url;
-}
