--- src/main/java/com/tourranger/image/travelAgency/repository/TravelAgencyRepositoryCustom.java
@@ -1,5 +1,0 @@
-package com.tourranger.image.travelAgency.repository;
-
-public interface TravelAgencyRepositoryCustom {
-
-}
