--- src/main/java/com/tourranger/image/travelAgency/repository/TravelAgencyRepositoryImpl.java
@@ -1,14 +1,0 @@
-package com.tourranger.image.travelAgency.repository;
-
-import org.springframework.stereotype.Repository;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-
-import lombok.RequiredArgsConstructor;
-
-@Repository
-@RequiredArgsConstructor
-public class TravelAgencyRepositoryImpl implements TravelAgencyRepositoryCustom {
-	private final JPAQueryFactory jpaQueryFactory;
-
-}
