+++ src/main/java/com/tourranger/thumbnailImage/repository/ThumbnailImageRepositoryCustom.java
@@ -1,0 +1,5 @@
+package com.tourranger.thumbnailImage.repository;
+
+public interface ThumbnailImageRepositoryCustom {
+
+}
