+++ src/main/java/com/tourranger/thumbnailImage/repository/ThumbnailImageRepositoryImpl.java
@@ -1,0 +1,14 @@
+package com.tourranger.thumbnailImage.repository;
+
+import org.springframework.stereotype.Repository;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+
+import lombok.RequiredArgsConstructor;
+
+@Repository
+@RequiredArgsConstructor
+public class ThumbnailImageRepositoryImpl implements ThumbnailImageRepositoryCustom {
+	private final JPAQueryFactory jpaQueryFactory;
+
+}
