+++ src/main/java/travelAgency/repository/TravelAgencyRepository.java
@@ -1,0 +1,10 @@
+package travelAgency.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import travelAgency.entity.TravelAgency;
+
+@Repository
+public interface TravelAgencyRepository extends JpaRepository<TravelAgency, Long>, TravelAgencyRepositoryCustom {
+}
