+++ src/main/java/travelAgency/repository/TravelAgencyRepositoryCustom.java
@@ -1,0 +1,5 @@
+package travelAgency.repository;
+
+public interface TravelAgencyRepositoryCustom {
+
+}
