+++ src/main/java/travelAgency/repository/TravelAgencyRepositoryImpl.java
@@ -1,0 +1,14 @@
+package travelAgency.repository;
+
+import org.springframework.stereotype.Repository;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+
+import lombok.RequiredArgsConstructor;
+
+@Repository
+@RequiredArgsConstructor
+public class TravelAgencyRepositoryImpl implements TravelAgencyRepositoryCustom {
+	private final JPAQueryFactory jpaQueryFactory;
+
+}
